Made modifications to the Departement model
- moved departement mode from core to org app
- removed dept_incharge field from departement
- removed all the mixins in lab app
