+++ core/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from django.contrib.auth.models import AbstractUser
 from django.utils.translation import gettext_lazy as _
 from . user_manager import UserManager
-from org.models import Org
+from org.models import Org, Department
     
 
 class User(AbstractUser):
@@ -20,6 +20,7 @@ class UserProfile(models.Model):
     first_name = models.CharField(max_length=255)
     last_name = models.CharField(max_length=255)
     org = models.ForeignKey(Org, on_delete=models.CASCADE)
+    dept = models.ForeignKey(Department, null=True, on_delete=models.SET_NULL)
     is_org_admin = models.BooleanField(_('is admin'), default=False)
     is_dept_incharge = models.BooleanField(_('is department incharge'), default=False)
     is_lab_staff = models.BooleanField(_('is lab staff'), default=False)
@@ -27,14 +28,6 @@ class UserProfile(models.Model):
     def __str__(self):
         return f"{str(self.first_name)} {str(self.last_name)}"
     
-    
-class Department(models.Model):
-    org = models.ForeignKey(Org, on_delete=models.CASCADE)
-    name = models.CharField(max_length=255)
-    incharge = models.OneToOneField(UserProfile, on_delete=models.DO_NOTHING)
-    created_on = models.DateTimeField(auto_now_add=True)
-    
-    def __str__(self):
-        return self.name
+
